Refactor(style): split component for style management
